refactor(config): 统一配置读取入口为 repo_settings
