+++ bot_worker.py
@@ -13,7 +13,7 @@ import os
 os.environ.setdefault("PALLAS_SHARD_ENABLED", "true")
 os.environ.setdefault("PALLAS_BOT_ROLE", "worker")
 
-from pallas.core.foundation.config.dotenv import apply_repo_settings_to_environ
+from pallas.core.foundation.config.repo_settings import apply_repo_settings_to_environ
 
 apply_repo_settings_to_environ()
